API update

+++ gizmo1b/library/libTec.h
@@ -7,13 +7,13 @@
 class LibTec
 {
 public:
-    typedef std::pair<int, double> TimeCurrent;
+    typedef std::pair<int, float> TimeCurrent;
     LibTec();
     virtual ~LibTec();
     void enable(bool en);
-    double getISense(); // Calculated Current Based on ADC voltage
-    double getVSense(); // Calculated Voltage Based on ADC voltage
-    void setCurrent(double value); // Calculated Closed Loop Current Output
+    float getISense(); // Calculated Current Based on ADC voltage
+    float getVSense(); // Calculated Voltage Based on ADC voltage
+    void setCurrent(float value); // Calculated Closed Loop Current Output
     // Linear Interpolation over time between set points,
     // Repeat output LOOPCOUNT times, Time (us)
     void setCurrent(int loopCount, std::vector<TimeCurrent>& setPoints);
